{"schema":"decodex/commit/1","summary":"Modularize host FFI capture frame","authority":"manual"}
